requirements.txt
asgiref==3.2.7
Django==3.0.5
djangorestframework==3.11.0
Pillow==7.1.1
pytz==2019.3
sqlparse==0.3.1
gunicorn
whitenoise
six

autopep8
pep8
boto3==1.10.49
django-storages
django-filter==2.2.0
psycopg2
django-heroku==0.3.1
django-stubs==1.4.0
djangorestframework-stubs==1.1.0
s3transfer==0.2.1